+++ tests/test_04_gemini.py
@@ -4,47 +4,21 @@
 import os
 import time
 from pathlib import Path
-import psutil
 import pytest
 from selenium.webdriver.common.by import By
 
-import talkingheads
+import generic
 from utils import get_driver_arguments
+from talkingheads import GeminiClient
 
 def test_start():
-    pytest.chathead = talkingheads.GeminiClient(**get_driver_arguments("gemini"))
+    pytest.chathead = GeminiClient(**get_driver_arguments("gemini"))
     assert pytest.chathead.ready, "The Client is not ready"
 
 
 def test_interaction():
-    response = pytest.chathead.interact(
-        "Without any explanation or extra information, just repeat the following: book."
-    )
-    assert (
-        "book" in response.lower()
-    ), f'response is not "book.", instead it responded {response}'
-    time.sleep(0.5)
+    return generic.test_interaction()
 
-
-# def test_regenerate():
-#     first_response = pytest.chathead.interact(
-#         "Without any explanation or extra information, type three animal names."
-#     ).lower()
-#     second_response = pytest.chathead.regenerate_response()
-#     assert first_response != second_response, "The regenerated response is the same."
-#     time.sleep(0.5)
-
-
-def test_interaction_with_file():
-    im_path = Path(os.path.dirname(__file__)) / 'test_assets/one_cat.jpg'
-    response = pytest.chathead.interact(
-        "Without further ado, just write 'indeed' if there is a cat in the uploaded image",
-        image_path=im_path
-    )
-    assert (
-        "indeed" in response.lower()
-    ), f"The word indeed doesn't exist in the response, instead it responded {response}"
-    time.sleep(0.5)
 
 def test_model():
     assert not pytest.chathead.switch_model("dream-model"), "Unexpected switch model behaviour"
@@ -72,9 +46,16 @@ def test_reset():
     ), "Chat is not empty"
     time.sleep(0.5)
 
+def test_interaction_with_file():
+    im_path = Path(os.path.dirname(__file__)) / 'test_assets/one_cat.jpg'
+    response = pytest.chathead.interact(
+        "Without further ado, just write 'indeed' if there is a cat in the uploaded image",
+        image_path=im_path
+    )
+    assert (
+        "indeed" in response.lower()
+    ), f"The word indeed doesn't exist in the response, instead it responded {response}"
+    time.sleep(0.5)
 
 def test_delete_chathead():
-    del pytest.chathead
-    assert not any(
-        "undetected_chromedriver" in p.name() for p in psutil.process_iter()
-    ), "Undetected chromedriver exists"
+    return generic.test_delete_chathead()